fix input argument module
