+++ spec/graphql/schema/relay_classic_mutation_spec.rb
@@ -4,15 +4,7 @@ require "spec_helper"
 describe GraphQL::Schema::RelayClassicMutation do
   describe ".input_object_class" do
     it "is inherited, with a default" do
-      custom_input = Class.new(GraphQL::Schema::InputObject)
-      mutation_base_class = Class.new(GraphQL::Schema::RelayClassicMutation) do
-        input_object_class(custom_input)
-      end
-      mutation_subclass = Class.new(mutation_base_class)
-
       assert_equal GraphQL::Schema::InputObject, GraphQL::Schema::RelayClassicMutation.input_object_class
-      assert_equal custom_input, mutation_base_class.input_object_class
-      assert_equal custom_input, mutation_subclass.input_object_class
     end
   end
 
@@ -26,15 +18,6 @@ describe GraphQL::Schema::RelayClassicMutation do
     end
   end
 
-  describe ".input_type" do
-    it "has a reference to the mutation" do
-      mutation = Class.new(GraphQL::Schema::RelayClassicMutation) do
-        graphql_name "Test"
-      end
-      assert_equal mutation, mutation.input_type.mutation
-    end
-  end
-
   describe ".null" do
     it "is inherited as true" do
       mutation = Class.new(GraphQL::Schema::RelayClassicMutation) do
@@ -42,17 +25,6 @@ describe GraphQL::Schema::RelayClassicMutation do
       end
 
       assert mutation.null
-    end
-  end
-
-  describe "input argument" do
-    it "sets a description for the input argument" do
-      mutation = Class.new(GraphQL::Schema::RelayClassicMutation) do
-        graphql_name "SomeMutation"
-      end
-
-      field = GraphQL::Schema::Field.new(name: "blah", resolver_class: mutation)
-      assert_equal "Parameters for SomeMutation", field.get_argument("input").description
     end
   end
